fix: prevent null reference in single player
On the addition of preventing the player from moving at the edge of the camera in duels, the Player script tries accessing otherPlayer even if its null, so now there is a check to see if otherPlayer is not null.
Also, deathManager is now not a raycast target so that the menu buttons can be pressed normally.